refactor(storage): consolidate BucketOptions and TransformOptions into Types.swift
